Internal re-architecting plus bug fixes.
(1) Adjusted the internal architecture to move more
functionality into the base class (base.py).  Much of
this functionality is implementation-independent.

(2) Made the exception-handling wrappers module-level
functions and enabled better interaction between the
wrapper in the base class and the one in the subclass.

(3) A few bug fixes on handling of extraArgs parameter
in the S3 subclass.